tez-ielts-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
loguru>=0.7.2
dishka>=1.1.0
sqlalchemy>=2.0.29
asyncpg>=0.29.0
aiosqlite>=0.20.0
alembic>=1.13.1
celery[redis]>=5.4.0
httpx>=0.27.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.9
slowapi>=0.1.9
email-validator>=2.1.1
pytest>=8.2.0
pytest-asyncio>=0.23.6
pytest-cov>=5.0.0
ruff>=0.4.4
bandit>=1.7.8
